# api-instagram/api-docs.json
{"swagger":"2.0","info":{"description":"Documentation Instagram api","version":"1.0.0","title":"Instagram API","contact":{"name":"clembnl","url":"https://github.com/clembnl","email":"cl.bonal@free.fr"},"license":{"name":"GNU gpl 3.0","url":"https://www.gnu.org/licenses/gpl-3.0.html"}},"host":"localhost:5000","basePath":"/","tags":[{"name":"commentary-controller","description":"Commentary Controller"},{"name":"like-controller","description":"Like Controller"},{"name":"post-controller","description":"Post Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/commentary/add":{"post":{"tags":["commentary-controller"],"summary":"addCom","operationId":"addComUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"commentaryDto","description":"commentaryDto","required":true,"schema":{"$ref":"#/definitions/CommentaryDto"}},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/commentary/update/{comID}":{"put":{"tags":["commentary-controller"],"summary":"updateCommentary","operationId":"updateCommentaryUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"comDto","description":"comDto","required":true,"schema":{"$ref":"#/definitions/CommentaryDto"}},{"name":"comID","in":"path","description":"comID","required":true,"type":"integer","format":"int32"},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/commentary/{comID}":{"delete":{"tags":["commentary-controller"],"summary":"deleteCommentary","operationId":"deleteCommentaryUsingDELETE","produces":["*/*"],"parameters":[{"name":"comID","in":"path","description":"comID","required":true,"type":"integer","format":"int32"},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/commentary/{postID}":{"get":{"tags":["commentary-controller"],"summary":"getComsByPost","operationId":"getComsByPostUsingGET","produces":["*/*"],"parameters":[{"name":"postID","in":"path","description":"postID","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/CommentaryDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/like/add":{"post":{"tags":["like-controller"],"summary":"addLike","operationId":"addLikeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"likeDto","description":"likeDto","required":true,"schema":{"$ref":"#/definitions/LikeDto"}},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/like/{likeID}":{"delete":{"tags":["like-controller"],"summary":"deleteLike","operationId":"deleteLikeUsingDELETE","produces":["*/*"],"parameters":[{"name":"likeID","in":"path","description":"likeID","required":true,"type":"integer","format":"int32"},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/like/{postID}":{"get":{"tags":["like-controller"],"summary":"getLikesByPost","operationId":"getLikesByPostUsingGET","produces":["*/*"],"parameters":[{"name":"postID","in":"path","description":"postID","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/LikeDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/post/":{"get":{"tags":["post-controller"],"summary":"getPosts","operationId":"getPostsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/PostDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/post/add":{"post":{"tags":["post-controller"],"summary":"addPost","operationId":"addPostUsingPOST","consumes":["multipart/form-data"],"produces":["*/*"],"parameters":[{"name":"image","in":"formData","description":"image","required":true,"type":"file"},{"in":"body","name":"postDto","description":"postDto","required":true,"schema":{"$ref":"#/definitions/PostDto"}},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/post/update/{postID}":{"put":{"tags":["post-controller"],"summary":"updatePost","operationId":"updatePostUsingPUT","consumes":["multipart/form-data"],"produces":["*/*"],"parameters":[{"name":"image","in":"formData","description":"image","required":false,"type":"file"},{"in":"body","name":"postDto","description":"postDto","required":true,"schema":{"$ref":"#/definitions/PostDto"}},{"name":"postID","in":"path","description":"postID","required":true,"type":"integer","format":"int32"},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/post/{postID}":{"delete":{"tags":["post-controller"],"summary":"deletePost","operationId":"deletePostUsingDELETE","produces":["*/*"],"parameters":[{"name":"postID","in":"path","description":"postID","required":true,"type":"integer","format":"int32"},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/post/{token}":{"get":{"tags":["post-controller"],"summary":"getPostsByToken","operationId":"getPostsByTokenUsingGET","produces":["*/*"],"parameters":[{"name":"token","in":"path","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/PostDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/user/all":{"get":{"tags":["user-controller"],"summary":"findAllUser","operationId":"findAllUserUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/User"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/user/signIn":{"post":{"tags":["user-controller"],"summary":"SignIn","operationId":"SignInUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"signInDto","description":"signInDto","required":true,"schema":{"$ref":"#/definitions/SignInDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SignInResponseDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/user/signup":{"post":{"tags":["user-controller"],"summary":"Signup","operationId":"SignupUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"signupDto","description":"signupDto","required":true,"schema":{"$ref":"#/definitions/SignUpDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"ApiResponse":{"type":"object","properties":{"message":{"type":"string"},"success":{"type":"boolean"},"timestamp":{"type":"string"}},"title":"ApiResponse"},"CommentaryDto":{"type":"object","properties":{"commentary":{"type":"string"},"id":{"type":"integer","format":"int32"},"postId":{"type":"integer","format":"int32"}},"title":"CommentaryDto"},"LikeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"}},"title":"LikeDto"},"PostDto":{"type":"object","properties":{"description":{"type":"string"},"id":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"}},"title":"PostDto"},"ResponseDto":{"type":"object","properties":{"message":{"type":"string"},"status":{"type":"string"}},"title":"ResponseDto"},"SignInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}},"title":"SignInDto"},"SignInResponseDto":{"type":"object","properties":{"status":{"type":"string"},"token":{"type":"string"}},"title":"SignInResponseDto"},"SignUpDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}},"title":"SignUpDto"},"User":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int32"},"password":{"type":"string"},"username":{"type":"string"}},"title":"User"}}}